requirements.txt
GitPython==3.1.31
numpy==1.23.5
openai==0.27.5
pymongo==4.4.1
python-dotenv==1.0.0
PyYAML==6.0
Requests==2.31.0
scikit_learn==1.3.0
SpeechRecognition==3.10.0
websockets==11.0.3
google-cloud-texttospeech